update main and manage file
